Fixed bug: databases.json list got overwritten on reinstall

+++ install.js
@@ -113,7 +113,9 @@ if (!fs.existsSync('root/content')) {
 // Overwrite ./private-workers directory contents
 qcd.sync(__dirname + '/static/private-workers', 'private-workers');
 // Create database list
-fs.writeFileSync('databases.json', '[]');
+if (!fs.existsSync('databases.json')) {
+    fs.writeFileSync('databases.json', '[]');
+}
 /* ===================
     3. Creation of databases
 =================== */
